Refactor services to use context api
