+++ src/App.js
@@ -1,9 +1,8 @@
 import './App.css';
-import { getContacts, getMessages } from './services/apiServices';
+import { getContacts, getMessages } from './services/Chat/chatContextBuilder';
 
 function App() {
 
-  getMessages()
 
   return (
     <div className="App">
