+++ src/services/Chat/ChatContext.js
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+const ChatContext = createContext()
+export default ChatContext
