+++ src/services/Chat/useChatContext.js
@@ -1,0 +1,5 @@
+import { useContext } from "react";
+import ChatContext from "./ChatContext";
+
+const useChatContext = () => useContext(ChatContext)
+export default useChatContext
